requirements.txt
Django==5.2.0
djangorestframework==3.16.0
django-cors-headers==4.7.0
mysqlclient==2.2.7
requests==2.31.0
cryptography==41.0.3
python-dotenv==1.0.0